Fix product features

+++ server/src/controllers/product-controller.js
@@ -20,13 +20,13 @@ class ProductController {
       condition,
       images: image_ids,
     });
-    await (await newProduct.save()).execPopulate({ path: 'images', select: 'filename filepath' });
+    await (
+      await newProduct.save()
+    ).execPopulate({ path: 'images', select: 'filename filepath type' });
 
     res.status(201);
     const result = {
-      data: {
-        product: newProduct,
-      },
+      data: newProduct,
       meta: {
         status: res.statusCode,
       },
@@ -35,15 +35,13 @@ class ProductController {
   });
 
   static getProduct = expressAsyncHandler(async (req, res) => {
-    const products = await Product.find().populate({
+    const productList = await Product.find().populate({
       path: 'images',
       select: 'filename filepath',
     });
 
     const result = {
-      data: {
-        products,
-      },
+      data: productList,
       meta: {
         status: res.statusCode,
       },
@@ -57,11 +55,10 @@ class ProductController {
       path: 'images',
       select: 'filename filepath',
     });
+    if (!product) throw createHttpError(404);
 
     const result = {
-      data: {
-        product,
-      },
+      data: product,
       meta: {
         status: res.statusCode,
       },
@@ -70,19 +67,30 @@ class ProductController {
   });
 
   static updateProduct = expressAsyncHandler(async (req, res) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      throw createHttpError(422, { errors: errors.array() });
+    }
     const { product_id } = req.params;
-    const product = await Product.findByIdAndUpdate(
+    const { name, price, stock, description, condition, image_ids } = req.body;
+
+    const updatedProduct = await Product.findByIdAndUpdate(
       product_id,
       {
-        ...req.body,
+        name,
+        price,
+        stock,
+        description,
+        condition,
+        images: image_ids,
       },
       { new: true },
     );
 
+    if (!updatedProduct) throw createHttpError(404);
+
     const result = {
-      data: {
-        product,
-      },
+      data: updatedProduct,
       meta: {
         status: res.statusCode,
       },
@@ -91,13 +99,17 @@ class ProductController {
   });
 
   static deleteProduct = expressAsyncHandler(async (req, res) => {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      throw createHttpError(422, { errors: errors.array() });
+    }
     const { product_id } = req.params;
-    const product = await Product.findByIdAndDelete(product_id);
+
+    const deletedProduct = await Product.findByIdAndDelete(product_id);
+    if (!deletedProduct) throw createHttpError(404);
 
     const result = {
-      data: {
-        product,
-      },
+      data: deletedProduct,
       meta: {
         status: res.statusCode,
       },
